// src/swagger/model/newPassword.ts
/**
 * WannaGo Client API
 * This is a sample Petstore server.  You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/). 
 *
 * OpenAPI spec version: 1.0.0
 * Contact: apiteam@swagger.io
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface NewPassword { 
    password?: string;
    emailAddress?: string;
    /**
     * mã xác thực nhận được từ email, sau khi yêu cầu đổi password
     */
    secret?: string;
}